refactor: cbor and cli
