+++ include/cbor/cb_cbor_string.h
