+++ src/cb_ops.c
@@ -15,8 +15,8 @@
 #include <limits.h>
 
 #include "cb_fs.h"
-#include "cb_cbor_int.h"
-#include "cb_cbor_simple.h"
+#include "cbor/cb_cbor_int.h"
+#include "cbor/cb_cbor_simple.h"
 #include "cb_ops.h"
 
 #define DATABASES_DIRECTORY "./databases"
