+++ src/cb_request_executor.c
@@ -9,9 +9,9 @@
 #endif
 
 
-#include "cb_cbor.h"
-#include "cb_cbor_int.h"
-#include "cb_cbor_string.h"
+#include "cbor/cb_cbor.h"
+#include "cbor/cb_cbor_int.h"
+#include "cbor/cb_cbor_string.h"
 #include "cb_msg.h"
 #include "cb_request.h"
 #include "cb_request_executor.h"
