+++ src/cbor/cb_cbor_string.c
@@ -5,9 +5,9 @@
 
 #include <string.h>
 
-#include "cb_cbor_string.h"
-#include "cb_cbor_int.h"
-#include "cb_cbor.h"
+#include "cbor/cb_cbor_string.h"
+#include "cbor/cb_cbor_int.h"
+#include "cbor/cb_cbor.h"
 
 //////////////
 // ENCODING //
