+++ src/cli/cb_cli_interpreter.c
@@ -3,10 +3,10 @@
  *
  */
 
-#include "cb_cli_interpreter.h"
+#include "cli/cb_cli_interpreter.h"
 #include "cb_msg_encode.h"
-#include "cb_cli_help.h"
-#include "cb_cbor_int.h"
+#include "cli/cb_cli_help.h"
+#include "cbor/cb_cbor_int.h"
 
 #include <errno.h>
 #include <stdio.h>
