+++ src/cli/cb_cli_scanner.c
@@ -3,7 +3,7 @@
  *
  */
 
-#include "cb_cli_scanner.h"
+#include "cli/cb_cli_scanner.h"
 
 scanner_t scanner_init(char *p_start, int p_len) {
   return (scanner_t){
