+++ tests/cbor/test_cb_cbor_string.c
@@ -8,7 +8,7 @@
 #include <stdlib.h>
 #include <string.h>
 
-#include "cb_cbor_string.h"
+#include "cbor/cb_cbor_string.h"
 
 void test_cb_cbor_encode_string() {
   // Size tests
