+++ tests/test_cb_ops.c
@@ -12,7 +12,7 @@
 #include <string.h>
 
 #include "cb_fs.h"
-#include "cb_cbor_int.h"
+#include "cbor/cb_cbor_int.h"
 #include "cb_ops.h"
 
 int fd_uint;
